Move speculative runner into own package
